requirements.txt
moondream
pillow